tst_qhttpserver: do not block the test with keychain access dialog
Our approach with a temporary keychain stopped working on macOS 15,
where we fortunately have a new option when importing a PKCS12 blob.
But it is possible to build using SDK 14 and then run on macOS 15,
where our trick would fail. In this case, skip the test/SSL usage.

Task-number: QTBUG-130500
Pick-to: 6.8
Change-Id: Iebe31d2d6affa686ccb4287cd80c2c5683cc787a
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>